fix(ci): maintainability line limit — extract hosted delegation config module
hosted_delegation.py exceeded the 1,500-line first-party file cap after recent
hosted validation identity changes; move config resolution into a sibling module
and re-export the public API from hosted_delegation for compatibility.
